popver common component setup
